+++ src/input_parsers/tests/Test_OnnxParser.h
@@ -154,6 +154,32 @@ class OnnxParserTestSuite : public CxxTest::TestSuite
         run_test( "transpose", input, output );
     }
 
+    void test_squeeze()
+    {
+        Vector<double> input = {
+            0, 1, -2,
+            0.5, 1.5, -3
+        };
+        Vector<double> output = {
+            0, 1, -2,
+            0.5, 1.5, -3
+        };
+        run_test( "squeeze", input, output );
+    }
+
+    void test_squeeze_with_axes()
+    {
+        Vector<double> input = {
+            0, 1, -2,
+            0.5, 1.5, -3
+        };
+        Vector<double> output = {
+            0, 1, -2,
+            0.5, 1.5, -3
+        };
+        run_test( "squeeze_with_axes", input, output );
+    }
+
     // Reference implementation for batch normalisation
     double batchnorm_fn(double input, double epsilon, double scale, double bias, double mean, double variance)
     {